safari-600.7-branch/LayoutTests/fast/events/popup-blocking-timers6-expected.txt

Click Here
Test calling window.open() with a 1001 ms delay. A popup should not be allowed.
PASS newWindow is undefined.
